session management: fix bugs (#47)
* enable proxy access

* temporary fix for server errors

* fix insert statement & enable deleting db tables

+++ classes/class.ilBigBlueButtonPlugin.php
@@ -18,11 +18,11 @@ class ilBigBlueButtonPlugin extends ilRepositoryObjectPlugin
     {
         global $ilDB;
         // removes plugin tables if they exist
-        /* if($ilDB->tableExists('rep_robj_xbbb_data'))
+         if($ilDB->tableExists('rep_robj_xbbb_data'))
          	$ilDB->dropTable('rep_robj_xbbb_data');
 
          if($ilDB->tableExists('rep_robj_xbbb_conf'))
-         	$ilDB->dropTable('rep_robj_xbbb_conf');*/
+         	$ilDB->dropTable('rep_robj_xbbb_conf');
         return true;
     }
 }
